Add new warning messages to gold files for implict event with selects.
An implicit event includes more than you may expect when selects are
added to the mix. Currently the select is ignored and the whole
structure word/array will be added to the sensitivity list. This can
create much confusion so warnings were added to let the user know
what is going on. This may be changed to only include what is
expected and if so the warnings will be removed at that time.