Update vcd-dup gold VCD file since the net order is different.
Development and V0.9 have different net orders so they produce
different, but functionally equivalent VCD files. This patch
gives each a VCD file. In the future we will have a real VCD
diff routine, but for now we use an external tool and then
update the VCD gold file as needed.