Add support for valgrind **\d+** lines.
Valgrind can print user input and when it does it uses **\d+**
before the output. We use this to display the number of signals
and nets that did not get deleted from the pool.

A search for files that have this problem was added to the
find_valg_errs script. It is currently commented out since I
have not started to look at cleaning these problems up and
there are a fair number of them.