Add p2951657.
Also touch up the expected behavior for pr377 and pr1002. The
time-0 X values are not really expected, and no other simulators
generate them. The fix for pr2951657 causes Icarus Verilog to
behave more like the rest of the world, so change the pr377 and
pr1002 gold files to reflect this new reality.