Fix invalid case5 test.
The original case5 test is invalid since the code incorrect drives
the output from both synthesized and non-synthesized code. Fixing
this allows V0.8 to synthesize this test correctly. V0.8 has also
been enhanced to print an error message and fail if this happens.

The vpi/memmon test was also marked NI for V0.8 since it fails
because of missing functionality that will not be added to V0.8.