Modify VPI tests to be more portable.
This patch fixes races in several of the VPI tests that led to different
output order when run on other simulators. It also adds a "vstd" category
to the VPI regression test list to cover tests where Icarus extends or
deviates from the standard and does a few other minor tweaks to make the
tests more portable.